Factor out getColonyClient function
