+++ src/clients/ColonyNetworkClient.ts
@@ -12,20 +12,8 @@ import {
   COLONY_NETWORK_ADDRESSES,
   REPUTATION_ORACLE_ENDPOINT,
 } from '../constants';
-import { ColonyVersion } from '../versions';
 import { IColonyNetwork, IColonyNetworkFactory, abis } from '../exports';
-import getColonyVersionClient from './Core/ColonyVersionClient';
-// FIXME: abstract this into own file
-import getColonyClientV1 from './Core/ColonyClientV1';
-import getColonyClientV2 from './Core/ColonyClientV2';
-import getColonyClientV3 from './Core/ColonyClientV3';
-import getColonyClientV4 from './Core/ColonyClientV4';
-import getColonyClientV5 from './Core/ColonyClientV5';
-import getColonyClientV6 from './Core/ColonyClientV6';
-import getColonyClientV7 from './Core/ColonyClientV7';
-import getColonyClientV8 from './Core/ColonyClientV8';
-import getColonyClientV9 from './Core/ColonyClientV9';
-import getTokenClient from './TokenClient';
+import { getColonyClient } from './Core/exports';
 import getTokenLockingClient, {
   TokenLockingClient,
 } from './TokenLockingClient';
@@ -163,110 +151,7 @@ const getColonyNetworkClient = (
       return getTokenLockingClient(tokenLockingAddress, signerOrProvider);
     };
 
-  networkClient.getColonyClient = async (
-    addressOrId: string | number,
-  ): Promise<AnyColonyClient> => {
-    let colonyAddress: string;
-    if (typeof addressOrId == 'number') {
-      colonyAddress = await networkClient.getColony(addressOrId);
-    } else {
-      colonyAddress = addressOrId;
-    }
-    const colonyVersionClient = getColonyVersionClient(
-      colonyAddress,
-      signerOrProvider,
-    );
-    // This is *kinda* hacky, but I have no better idea ¯\_(ツ)_/¯
-    // We have to get the version somehow before instantiating the right contract version
-    const versionBN = await colonyVersionClient.version();
-    const version = versionBN.toNumber() as ColonyVersion;
-    let colonyClient: AnyColonyClient;
-    switch (version) {
-      case 1: {
-        colonyClient = getColonyClientV1.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 2: {
-        colonyClient = getColonyClientV2.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 3: {
-        colonyClient = getColonyClientV3.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 4: {
-        colonyClient = getColonyClientV4.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 5: {
-        colonyClient = getColonyClientV5.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 6: {
-        colonyClient = getColonyClientV6.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 7: {
-        colonyClient = getColonyClientV7.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 8: {
-        colonyClient = getColonyClientV8.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      case 9: {
-        colonyClient = getColonyClientV9.call(
-          networkClient,
-          colonyAddress,
-          signerOrProvider,
-        );
-        break;
-      }
-      default: {
-        throw new Error('Colony version not supported');
-      }
-    }
-
-    const tokenAddress = await colonyClient.getToken();
-    colonyClient.tokenClient = await getTokenClient(
-      tokenAddress,
-      signerOrProvider,
-    );
-
-    return colonyClient;
-  };
+  networkClient.getColonyClient = getColonyClient.bind(networkClient);
 
   networkClient.getMetaColonyClient = async (): Promise<AnyColonyClient> => {
     const metaColonyAddress = await networkClient.getMetaColony();
