+++ src/clients/Core/exports.ts
@@ -1,12 +1,18 @@
-import { ColonyClientV1 } from './ColonyClientV1';
-import { ColonyClientV2 } from './ColonyClientV2';
-import { ColonyClientV3 } from './ColonyClientV3';
-import { ColonyClientV4 } from './ColonyClientV4';
-import { ColonyClientV5 } from './ColonyClientV5';
-import { ColonyClientV6 } from './ColonyClientV6';
-import { ColonyClientV7 } from './ColonyClientV7';
-import { ColonyClientV8 } from './ColonyClientV8';
-import { ColonyClientV9 } from './ColonyClientV9';
+import { ColonyNetworkClient } from '../ColonyNetworkClient';
+import getColonyVersionClient from './ColonyVersionClient';
+import getTokenClient from '../TokenClient';
+
+import getColonyClientV1, { ColonyClientV1 } from './ColonyClientV1';
+import getColonyClientV2, { ColonyClientV2 } from './ColonyClientV2';
+import getColonyClientV3, { ColonyClientV3 } from './ColonyClientV3';
+import getColonyClientV4, { ColonyClientV4 } from './ColonyClientV4';
+import getColonyClientV5, { ColonyClientV5 } from './ColonyClientV5';
+import getColonyClientV6, { ColonyClientV6 } from './ColonyClientV6';
+import getColonyClientV7, { ColonyClientV7 } from './ColonyClientV7';
+import getColonyClientV8, { ColonyClientV8 } from './ColonyClientV8';
+import getColonyClientV9, { ColonyClientV9 } from './ColonyClientV9';
+
+export type ColonyVersion = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
 
 export { ColonyClientV1 } from './ColonyClientV1';
 export { ColonyClientV2 } from './ColonyClientV2';
@@ -18,11 +24,6 @@ export { ColonyClientV7 } from './ColonyClientV7';
 export { ColonyClientV8 } from './ColonyClientV8';
 export { ColonyClientV9 } from './ColonyClientV9';
 
-// FIXME: use this pattern
-// import getWhitelistClientV1, {
-//   WhitelistClient as WhitelistClientV1,
-// } from './WhitelistClientV1';
-
 export type AnyColonyClient =
   | ColonyClientV1
   | ColonyClientV2
@@ -33,3 +34,110 @@ export type AnyColonyClient =
   | ColonyClientV7
   | ColonyClientV8
   | ColonyClientV9;
+
+export async function getColonyClient(
+  this: ColonyNetworkClient,
+  addressOrId: string | number,
+): Promise<AnyColonyClient> {
+  let colonyAddress: string;
+  const signerOrProvider = this.signer || this.provider;
+  if (typeof addressOrId == 'number') {
+    colonyAddress = await this.getColony(addressOrId);
+  } else {
+    colonyAddress = addressOrId;
+  }
+  const colonyVersionClient = getColonyVersionClient(
+    colonyAddress,
+    signerOrProvider,
+  );
+  // This is *kinda* hacky, but I have no better idea ¯\_(ツ)_/¯
+  // We have to get the version somehow before instantiating the right contract version
+  const versionBN = await colonyVersionClient.version();
+  const version = versionBN.toNumber() as ColonyVersion;
+  let colonyClient: AnyColonyClient;
+  switch (version) {
+    case 1: {
+      colonyClient = getColonyClientV1.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 2: {
+      colonyClient = getColonyClientV2.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 3: {
+      colonyClient = getColonyClientV3.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 4: {
+      colonyClient = getColonyClientV4.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 5: {
+      colonyClient = getColonyClientV5.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 6: {
+      colonyClient = getColonyClientV6.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 7: {
+      colonyClient = getColonyClientV7.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 8: {
+      colonyClient = getColonyClientV8.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    case 9: {
+      colonyClient = getColonyClientV9.call(
+        this,
+        colonyAddress,
+        signerOrProvider,
+      );
+      break;
+    }
+    default: {
+      throw new Error('Colony version not supported');
+    }
+  }
+
+  const tokenAddress = await colonyClient.getToken();
+  colonyClient.tokenClient = await getTokenClient(
+    tokenAddress,
+    signerOrProvider,
+  );
+
+  return colonyClient;
+}
